refactor(LottoGameReq, LottoResultRes): 디렉토리 구조 변경 및 클래스 이름 변경
