+++ src/main/java/lotto/domain/lotto/dto/request/LottoGameReq.java
@@ -1,14 +1,14 @@
-package lotto.domain.lotto.dto;
+package lotto.domain.lotto.dto.request;
 
 import java.util.List;
 
-public class LottoReq {
+public class LottoGameReq {
     private final List<List<Integer>> purchasedNumbers;
     private final List<Integer> winningNumbers;
     private final int bonusNumber;
     private final int cost;
 
-    private LottoReq(
+    private LottoGameReq(
             List<List<Integer>> purchasedNumbers,
             List<Integer> winningNumbers,
             int bonusNumber,
@@ -20,13 +20,13 @@ public class LottoReq {
         this.cost = cost;
     }
 
-    public static LottoReq of(
+    public static LottoGameReq of(
             List<List<Integer>> purchasedNumbers,
             List<Integer> winningNumbers,
             int bonusNumber,
             int cost
     ) {
-        return new LottoReq(purchasedNumbers, winningNumbers, bonusNumber, cost);
+        return new LottoGameReq(purchasedNumbers, winningNumbers, bonusNumber, cost);
     }
 
     public List<List<Integer>> getPurchasedNumbers() {
