+++ src/main/java/lotto/domain/lotto/service/LottoService.java
@@ -3,6 +3,4 @@ package lotto.domain.lotto.service;
 public interface LottoService {
 
 
-
-
 }
